requirements.txt
numpy<2.0
pandas==2.1.4
PyYAML==6.0.1
uproot==5.2.0
boost_histogram
sklearn